requirements.txt
# BikeLogic - Requisitos del Proyecto
# Instalación: pip install -r requirements.txt

# Google APIs
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0

# Geocodificación y mapas
requests>=2.28.0
geopy>=2.3.0

# Procesamiento geométrico
shapely>=2.0.0

# Optimización de rutas
ortools>=9.5.0

# Cálculos numéricos
numpy>=1.24.0
scipy>=1.10.0
